requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astroid==2.3.3
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-compat==1.0.15
djangorestframework==3.12.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.1.2
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
install==1.3.4
isort==4.3.21
joblib==1.0.1
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mbed-cloud-sdk==2.2.0
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.2
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.59.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
